+++ lib/plugins/index.js
@@ -9,7 +9,7 @@ module.exports = [
   require('./deploy.js'),
   require('./invoke.js'),
   require('./info.js'),
-  require('./logs/logs.js'),
+  require('./logs.js'),
   require('./metrics/metrics.js'),
   require('./print/print.js'),
   require('./remove/remove.js'),
